Today's Commit Includes Solution For Test Case Day 18
Which Has Two Distinct Tasks To Achieve, One Of Them Is To Validate Log
Out And Another One Is To Create And Validate A New Customer.
On Added note Today's Commit Is The Final Commit On Guru99 Demo Bank
Series Project, A Very Fine Project for Upcoming Testers With Hand On
Test Guidelines And Scenarios To Understand And Develop Test Solution
Better.